PORTAL-258: Removed News Detail page
Home page
Update the Home page so that the Latest Updates 'Read More >' links go to that topic on the news main page not a details page

News Main page
Update the news items so that they no longer link to a details page.

News Details page
Removed all of these pages.