PORTAL-347: Design & Implement - Carousel Pause Button
The home page include includes a pause indicator that can be used to pause the automatic advancement of the carousel
On click, the indicator stops the advancement of the carousel.  On second click the carousel re-starts automatic advancement
If paused, on page refresh the carousel restarts.
When paused, the up and down carousel navigation arrow still work.  Using the navigation arrows does not restart advancement
Keyboard users should be able to navigate to and use the pause indicator.