PORTAL-2272:  Design QA fixes - Study Details Page (#724)
* Replace non-508 compliant color with #00818D (updated on design- prev was #0095A2)

Changed the .studyItemTitle color from #0095A2 to #00818D in the StudiesDetailBodyContainer styled component for improved visual consistency.

* Height of banner should be 64px instead of 68px

Updated the padding for the .studyDetailText class in StudiesDetailContainer to improve visual alignment and spacing.

* Breadcrumb should be aligned with Nav

Replaces margin-left with padding-left in the study detail container for better spacing and updates the breadcrumb width for large screens from 1350px to 1420px to improve layout consistency.